+++ static/js/detail.js
@@ -52,10 +52,9 @@ function handle_delete(event) {
     });
 }
 
-const promise_charge = fetch(END_POINT + "/charge",{
+const promise_charge = fetch(END_POINT + "/charge", {
   method: "GET",
   headers: { Authorization: "Bearer " + get_token() },
-
 })
   .then(response => response.json()) // 将响应转换为JSON
   .then(data => {
@@ -176,9 +175,9 @@ document.addEventListener("DOMContentLoaded", () => {
     form_api["type_id"] = parseInt(form_api["type_id"], 10);
     form_api["charge_id"] = parseInt(form_api["charge_id"], 10);
     form_api["status_id"] = parseInt(form_api["status_id"], 10);
-    form_api["balance_payment"] = parseInt(form_api["balance_payment"], 10);
-    form_api["payment"] = parseInt(form_api["payment"], 10);
-    form_api["cost"] = parseInt(form_api["cost"], 10);
+    form_api["balance_payment"] = parseFloat(form_api["balance_payment"], 10);
+    form_api["payment"] = parseFloat(form_api["payment"], 10);
+    form_api["cost"] = parseFloat(form_api["cost"], 10);
     var checkboxes = document.getElementsByName("m");
     var selectedm = [];
